requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.3
django-debug-toolbar==4.4.6
django-filebrowser-no-grappelli==4.0.2
django-tinymce==4.1.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
isort==5.13.2
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
mypy-extensions==1.0.0
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
ruff==0.8.0
six==1.17.0
sqlparse==0.5.2
transliterate==1.10.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.2
